Move difficulty estimator from int to long so that it can properly represent > 2.1B clauses

+++ include/stp/STPManager/STP.h
@@ -121,8 +121,8 @@ public:
   // calls sizeReducing and the bitblasting simplification.
   ASTNode callSizeReducing(ASTNode simplified_solved_InputToSAT,
                            BVSolver* bvSolver, PropagateEqualities* pe,
-                           const int initial_difficulty_score,
-                           int& actualBBSize);
+                           const long initial_difficulty_score,
+                           long& actualBBSize);
 
   void ClearAllTables(void)
   {
